Forced fit data to float data type
I ran into trouble with the fit data being an integer data type when I was testing the simplex method of optimization on families of simulations, so I thought that I'd go ahead and force the fit data to be of float data type.

+++ worker_family.py
@@ -90,7 +90,7 @@ class WorkerFamily(SimplexPlotterMixin, ParameterManipulatorMixin):
     self.setup_parameters()
     self.read_target_data()
 
-    self.fit_data = self.target_data[:, 1].copy()
+    self.fit_data = self.target_data[:, 1].copy().astype(np.float)
 
     child_dict = {
       "fibersim_file":self.fibersim_file,
